+++ src/plugins/store/reducers/user/workflow.ts
@@ -1,0 +1,42 @@
+import { userThunks, userActions } from '@plugins/store';
+import { sessionActions } from '@plugins/store';
+import { IUserLoginInput, IUserRegisterInput } from '@plugins/backend/types';
+
+import type { AppDispatch } from '@plugins/store';
+import type { NavigateFunction } from 'react-router-dom';
+
+export const userWorkflows = {
+  login: (dispatch: AppDispatch, arg: IUserLoginInput) => {
+    dispatch(userActions.setLoginType('login'));
+    dispatch(userThunks.loginUser(arg))
+      .unwrap()
+      .then((userAuth) =>
+        dispatch(
+          userThunks.verifyUser({
+            token: userAuth.token,
+          }),
+        ).unwrap(),
+      )
+      .then(() => dispatch(sessionActions.setUserVerifyStatus(true)))
+      .catch(() => dispatch(userActions.signOut()));
+  },
+  register: (dispatch: AppDispatch, arg: IUserRegisterInput) => {
+    dispatch(userActions.setLoginType('register'));
+    dispatch(userThunks.registerUser(arg))
+      .unwrap()
+      .then((userAuth) =>
+        dispatch(
+          userThunks.verifyUser({
+            token: userAuth.token,
+          }),
+        ).unwrap(),
+      )
+      .then(() => dispatch(sessionActions.setUserVerifyStatus(true)))
+      .catch(() => dispatch(userActions.signOut()));
+  },
+  signOut: (dispatch: AppDispatch, navigate: NavigateFunction) => {
+    dispatch(userActions.signOut());
+    dispatch(sessionActions.setUserVerifyStatus(false));
+    navigate('/auth');
+  },
+};