+++ src/routes/index.tsx
@@ -1,3 +1,4 @@
 export * from './Root';
 export * from './Home';
+export * from './auth';
 export * from './templates';